Fixed rnamigos structure, fixed library option, fixed imports, fixed endpoints
