+++ core/src/nx2fp.py
@@ -8,7 +8,7 @@ import torch
 import dgl
 import itertools
 
-from rgcn import Model
+from .rgcn import Model
 
 nuc_map = {n:i for i,n in enumerate(['A', 'C', 'G', 'N', 'U'])}
 
@@ -42,13 +42,11 @@ def load_model(run):
         Load full trained model with id `run`
 
     """
-
-    meta = pickle.load(open(f'static/models/{run}/meta.p', 'rb'))
+    meta = pickle.load(open(f'core/static/models/{run}/meta.p', 'rb'))
 
     edge_map = meta['edge_map']
     num_edge_types = len(edge_map)
-
-    model_dict = torch.load(f'static/models/{run}/{run}.pth', map_location='cpu')
+    model_dict = torch.load(f'core/static/models/{run}/{run}.pth', map_location='cpu')
     model = Model(dims=meta['embedding_dims'], attributor_dims=meta['attributor_dims'], num_rels=num_edge_types,
                   num_bases=-1,
                   device='cpu',
