+++ core/static/libraries/pdb_rna_smiles.txt
